+++ src/main/java/com/ubb/eventapp/controller/GroupController.java
@@ -27,14 +27,14 @@ public class GroupController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Group> findById(@PathVariable Integer id) {
+    public ResponseEntity<Group> findById(@PathVariable Long id) {
         return service.findById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @GetMapping("/representative/{userId}")
-    public ResponseEntity<List<Group>> findByRepresentative(@PathVariable Integer userId) {
+    public ResponseEntity<List<Group>> findByRepresentative(@PathVariable Long userId) {
         return ResponseEntity.ok(service.findByRepresentativeUser(userId));
     }
