+++ src/main/java/com/ubb/eventapp/controller/RegistrationController.java
@@ -25,7 +25,7 @@ public class RegistrationController {
     }
 
     @GetMapping("/{eventId}/{userId}")
-    public ResponseEntity<Registration> findById(@PathVariable Integer eventId, @PathVariable Integer userId) {
+    public ResponseEntity<Registration> findById(@PathVariable Long eventId, @PathVariable Long userId) {
         RegistrationId id = new RegistrationId(eventId, userId);
         return service.findById(id)
                 .map(ResponseEntity::ok)
