+++ src/main/java/com/ubb/eventapp/repository/SupportTicketRepository.java
@@ -3,5 +3,5 @@ package com.ubb.eventapp.repository;
 import com.ubb.eventapp.model.SupportTicket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SupportTicketRepository extends JpaRepository<SupportTicket, Integer> {
+public interface SupportTicketRepository extends JpaRepository<SupportTicket, Long> {
 }
