+++ src/main/java/com/ubb/eventapp/repository/TokenRepository.java
@@ -12,7 +12,7 @@ public interface TokenRepository extends JpaRepository<Token, Long> {
             select t from Token t inner join t.user u
             where u.id = :userId and (t.expired = false or t.revoked = false)
             """)
-    List<Token> findAllValidTokenByUser(Integer userId);
+    List<Token> findAllValidTokenByUser(Long userId);
 
     Optional<Token> findByAccessToken(String token);
 }
